Refactor: Move archiveComponent to store actions and remove use-archive hook
- Moved the `archiveComponent` function from the `use-archive` hook to `store/actions.ts` for better organization.
- Deleted the now redundant `use-archive.ts` file.
- Updated imports in `components-item.tsx` to reflect the new location of `archiveComponent`.
- Enhanced state management by integrating document modification status directly within the action.

This change improves code structure and maintains consistency in component archiving functionality.
