+++ packages/packages/core/editor-components/src/components/components-tab/components-item.tsx
@@ -17,7 +17,7 @@ import {
 } from '@elementor/ui';
 import { __ } from '@wordpress/i18n';
 
-import { archiveComponent } from '../../hooks/use-archive';
+import { archiveComponent } from '../../store/actions';
 import { loadComponentsAssets } from '../../store/load-components-assets';
 import { type Component } from '../../types';
 import { getContainerForNewElement } from '../../utils/get-container-for-new-element';
