+++ packages/packages/core/editor-components/src/store/actions.ts
@@ -1,4 +1,4 @@
-import { type V1Document } from '@elementor/editor-documents';
+import { setDocumentModifiedStatus, type V1Document } from '@elementor/editor-documents';
 import { __getStore as getStore } from '@elementor/store';
 
 import { type ComponentsPathItem, slice } from './store';
@@ -19,3 +19,15 @@ export function updateCurrentComponent( {
 	dispatch( slice.actions.setPath( path ) );
 	dispatch( slice.actions.setCurrentComponentId( currentComponentId ) );
 }
+
+export const archiveComponent = ( componentId: number ) => {
+	const store = getStore();
+	const dispatch = store?.dispatch;
+
+	if ( ! dispatch ) {
+		return;
+	}
+
+	dispatch( slice.actions.archive( componentId ) );
+	setDocumentModifiedStatus( true );
+};
